pmb.config: combine files and add accessor
simplify the file heirarchy by combining the load, save, and
merge_with_args single-function files into config.py.

At the same time, save the loaded config in a global with runtime
overrides for arguments from the cmdline and add a get() method for
fetching config options as an alternative to accessing the args
variable.

Signed-off-by: Caleb Connolly <kc@postmarketos.org>