pmb.core: strictly validate chroot suffix
Perform a full validation on the suffix, this encodes the rules on
suffix naming and will cause a runtime exception if a suffix doesn't
follow the rules.

The rules on how a chroot suffix is named are well-defined and followed
consistently throughout the codebase, so this should have no breaking
changes (unless I made a mistake migrating everything to the new suffix
type).