pmb.build: add a bootstrap build mode
Some packages have cyclical dependencies and benefit from having a
separate bootstrap variant, make this process easier by automatically
setting up an environment similar to what's used in the openssh alpine
package.

Introduce a new non-standard APKBUILD variable "_bootstrap_exclude"
which is used to specify dependencies which can be skipped when
bootstrapping, allowing the dependency cycle to be broken.

In addition, set "BOOTSTRAP=1" in the environment for abuild. The
APKBUILD can then change behaviour based on this variable.

This makes bootstrapping a bunch of new packages much easier, and lets
us leave the bootstrapping capabilities in the source repo so that if
need be they can be bootstrapped again from scratch.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>