pmb.build: ensure all specified packages are built with --force
When running build --force, if a later package is a dependency of an
earlier package, it will be "visited" by the dependency resolver,
erroneously causing it to be skipped when later in the session it's
visited to force build.

Fix this by overriding skip_already_build() if force is specified.

Signed-off-by: Caleb Connolly <kc@postmarketos.org>