pmb.build: add a dirty-build feature for local development
With automated checksums, the last piece of the puzzle for quick
development of local packages is to fix the pkgver automatically too.
Add a new --dirty flag to pmbootstrap build which will append the current
datetime to the pkgver. This has two main purposes:
1. Ensures that the newly built package will get picked up by apk
2. Makes it clear that the package is from a local "dirty" build and not
from any binary repo.

The dirty flag only applies when either the local checksums are out of
date, or if the local package source dir has been modified since the
last build. This ensures that only packages the user is working on will
get built when this flag is set.

With this logic in place, it is now sensible to add a '--build' flag to
pmbootstrap sideload, the behaves pretty much like a shortcut for
"pmbootstrap build --dirty foobar && pmbootstrap sideload foobar", just
combined for speed. A flag like this didn't make much sense before as
local modification necessitated fixing the checksum and bumping the
pkg{ver,rel}.

Signed-off-by: Caleb Connolly <kc@postmarketos.org>