pmb.install: clean up initialization of recommends/selected
Avoid passing in a boolean to decide if this is the first time the
function has been called and instead just initialise the globals at the
top of the file.

We can figure out the state management if/when we want
to handle doing multiple installs in a single invocation of pmb.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>