WIP: handle extra-repos as independent source repos
To facilitate nicely integrating systemd packages, we should treat them
as a separate source and binary repo. We have the binary repo part
mostly done, however we ideally want to have the extra-repos/systemd dir
in pmaports treated as a separate source repo. In this way, we can have
the openrc version of a package like gnome-shell-mobile in pmaports, and
then the systemd variant in the systemd repo.

Usually this isn't allowed, but by teaching pmbootstrap to handle
multiple package source repos, we can do some trickery to mostly make it
work.

FIXME: we need to find a way to know if the systemd repo should be
parsed or not, and this currently isn't possible due to a cyclical
dependency during init.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>