pmb.chroot.zap: don't calculate size (MR 2252)
It's nice to know how much space gets free'd when zapping, but I often
find myself with a lot of chroot's and other junk, and on my laptop
running "du" across all of this takes quite a few seconds. As this is
purely cosmetic, it doesn't justify taking such a long time.

Remove the size calculation code to substantially speed up zap.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>