pkgrel_bump --auto: ignore -service depends (MR 2321)
Fix the "upstream compat" test that is currently failing with:
  Increase 'postmarketos-ui-phosh' pkgrel (6 -> 7), missing depend(s): phosh-service

The mentioned dependency is in the postmarketos-ui-phosh-systemd
subpackage, and the phosh-service package is intentionally not provided
by anything in pmaports master right now. It will be there once we merge
the master_staging_systemd branch with master, which requires
pmbootstrap and bpo changes first.

Make the "upstream compat" test useful again by not failing on this for
now.

This patch is for the 2.3.x branch of pmbootstrap, which is used by CI
while we prepare for pmbootstrap v3.

Fixes: montoring issue 8