pmb.chroot.init: build chroots with "user repo" enabled
This makes sure that chroots are build with the local "user repo"
enabled. It fixes situations where a chroot is built with an older
version of something used to init a chroot, like alpine-base.
The comment @ pmb.chroot.init.py:187 suggests that folks expected it to
work this way all along, since init will explicitly call `apk upgrade
-a` to pull in any local repo updates.