pmb.sideload: Run $ uname -m in subshell (MR 2390)
Some shells, in particular Nushell, replace the built-in
POSIX-compatible uname with a custom one that does not support common
flags. I opened an issue in the Nushell repository about this, but it
was rejected on the grounds that POSIX compatibility is not a goal. As
such, run this in a subshell to ensure that we get the expected uname
interface.

Nushell issue: https://github.com/nushell/nushell/issues/12570