Configure ruff to lint imports
Rather than manually check that we're following conventions for imports
and to avoid developers manually having to tinker formatting, let ruff
do checking and fixing of import style.

We have a bit of a mix in the codebase right now; ruff's default
settings produce the smallest diff, so it could be said to be the
closest to what's being done right now. The second smallest diff is
achieved with a single line per import.