pmb.helpers.package.get: fix !ret + !ret_repo case
"ret" gets initialized as {} at the beginning of the function. Fix the
code path where 'ret_repo = pmb.parse.apkindex.package' may return None,
in which case we should not overwrite "ret" with None.

Otherwise we suddenly change the datatype from dict to None and
subsequent code fails.

Fixing this uncovered another bug, the arch compare doesn't work because
of invalid datatype. Add a workaround for it to not risk further fallout
(pmb#2455).