helpers.pmaports.find: fix subpackages logic
Fix that pmbootstrap can't find the dependency "musl-dev-aarch64" of
"gcc-aarch64" in pmaports.

If we can't find the package by the full name in pmaports (in this case,
musl-dev-aarch64 is provided by musl-aarch64), then we make a guess of
what the main pkgname could be (as optimization, e.g. subpkgname-dev ->
subpkgname), and then verify that guess.

However for musl-dev-aarch64 the guess does not work as it does not
follow the usual packaging scheme, and so we must iterate through all
APKBUILDs in pmaports and check if any of these provide musl-dev-aarch64
in subpackages.

Without this patch, going through all APKBUILDs only happens if we have
made any guess before (which is None for musl-dev-aarch64).