code dump integrate mkosi-sandbox
Vendor mkosi-sandbox and beging The Refactoring...

Since almost all of the mount calls we do are for bindmounts, we can let
mkosi-sandbox handle it all instead of actually calling "sudo mount", we
can also use sandbox to protect the host system from dangers by running
things like "apk" inside it where it has no ability to write to anything
on the host (and can't even read /home, /var, or /etc).

Actually stuffing the rootfs into an image is where it gets harder. We
can do mkfs.ext4 quite easily by passing it "-d /chroot_rootfs...", and
use mcopy to do the same for the fat32 boot partition, but this is
harder for btrfs and other filesystems (maybe not??).

The biggest issue currently is that we use loop mounting to handle the
partition table. To make this work right now there's a workaround to
chown the loop device as the user running pmbootstrap. This is pretty
dangerous though, we will need a better approach!

All in, a full "pmbootstrap install" now takes 8 sudo calls, but i think
that could be easily bought down.

Issues/ideas:

* Only ext4 works, other FSs aren't implemented, rsync is broken, lots
  of other stuff is probably also broken...
* pmbootstrap build is broken
* Since mkosi-sandbox is quite heavy and we have to do a lot of argument
  building, it would be good to reduce the number of calls. We could
  look at "building" chunks of commands to run (a script if you will)
  and running them all at once which would drastically speed things up
* loop mounting security/permissions issue mentioned above (plus most of
  the sudo calls are for this). Maybe we could avoid loop mounting and
  do crazy dd offsets instead? systemd-repart would solve this also...
* Using file-backed namespaces might(?) let us simplify some things,
  mkosi-sandbox would also need to gain support though. e.g. we could
  call mount with --map-users or something. idk if this could be
  helpful.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>