install: support rootfs-as-initramfs
Implement support for stuffing the entire rootfs into an initramfs
archive. This is useful in some cases for testing.

running "pmbootstrap install --initramfs" will output a regular
postmarketOS boot partition, except the initramfs IS the root partition.

A cut down "none" UI install is still ~250mb with firmware and systemd.

Automatically install and enable the new postmarketos-network-gadget
package so that the USB network gadget will be set up during boot as
normal.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>