core: chroot: add a sandbox flag
A sandbox-chroot is identical except for using the "sandbox_" prefix
instead of the usual "chroot_" prefix. Due to how user namespacing works
it isn't possible to mix and match accessing chroots from the sandbox or
the usual chroot() method since the permissions won't match up.

This allows re-using the chroot API for sandboxed chroots.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>