core: sandbox: introduce experimental sandbox utilities
Introduce an initial API for running commands in a user namespace by
calling out to mkosi-sandbox (which is vendored in pmbootstrap).

There are two types of Sandbox currently available:

1. The HostSandbox() sets up a user namespace using your hosts
   /usr in read-only mode. This is useful for running apk-tools for
   example, since it ensures you will never pollute your host rootfs.
2. The ChrootSandbox() sets up a user namespace inside a pmOS rootfs, it
   does this by using the HostSandbox() (indirectly via ApkTools) to
   create a rootfs, then enters it via mkosi sandbox.

The ApkTools() class serves as an example user, the API is just a
placeholder for now, but it works using contextlib to "build" the host
sandbox configuration (effectively a chain of arguments to
mkosi-sandbox) and then wraps standard apk commands through the sandbox.

Properly integrating user namespacing into pmbootstrap will be a fairly
significant challenge due to how fundamentally different it is to the
current chroot method. The hope is to begin by using the new Sandbox API
for more "atomic" operations like updating checksums, analyzing boot
images, etc, then eventually for building packages and lastly for
generating images.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>