commands: introduce sandbox command
Introduce a new sandbox command, this will drop the user into an
interactive shell inside a user namespace, either in the context of
their host or a persistent native Alpine chroot.

The current working directory will also be made available inside the
sandbox, this offers a few practical usecases already such as being able
to test building a piece of software for Alpine when your host machine
is running a different distro.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>