parse: apkbuild: rewrite parser to be dynamic
The current static parser is limited in functionality, since it doesn't
run the APKBUILD it can't handle conditional logic like bootstrap
dependencies.

Let's solve this by replacing it with a new parser, written in posix SH
which exports the APKBUILD data in JSON format.

This new parser is capable of parsing all APKBUILD files in aports.
However for the parser to be safe it must run in a sandbox, introducing
quite a lot of overhead.

This is partially addressed by caching all parsed APKBUILD files in JSON
format, but could be further improved by allowing for multiple APKBUILD
files to be parsed with a single invocation.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>