ci/markdown: fix shellcheck errors (MR 2485)
Don't try to modify a variable in the while loop, as it runs in a
subshell because of the pipe. Fail on the first error instead, like we
have it in .ci/shellcheck.sh.

This can be optimized later on (for both scripts) if we really want to.

Fix for:

  In ./.ci/markdown.sh line 17:
  	markdownlint-cli "$file" || MDL_FAILED=1
                                      ^--------^ SC2030 (info): Modification of MDL_FAILED is local (to subshell caused by pipeline).
  In ./.ci/markdown.sh line 20:
  if [ "$MDL_FAILED" = "1" ]; then
        ^---------^ SC2031 (info): MDL_FAILED was modified in a subshell. That change might be lost.