docs/cross_compiling: fix makedeps host/build desc
* makedepends_build is for the native chroot
* makedepends_host is for the foreign chroot

I had it the other way around in the documentation. The variables are so
far not documented here yet:
* https://gitlab.alpinelinux.org/alpine/abuild/-/blob/master/APKBUILD.5.scd
* https://wiki.alpinelinux.org/wiki/APKBUILD_Reference

But it now matches the comments Caleb had written for the cross-native2
implementation, and also what can be found in e.g. the abuild APKBUILD
in aports.git:

  makedepends_build="pkgconfig scdoc"
  makedepends_host="openssl-dev>3 zlib-dev"

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2566
[ci:skip-build]: already built successfully in CI