remove ondev support
The on-device installer has been unsupported for a while, start removing
it from pmbootstrap to simplify the partitioning code, for example we no
longer need to support having some "reserved" space in the rootfs.

This follows various discussions which essentially indicated that future
on device installation methods will take a different approach which will
not require bespoke partitioning.

Signed-off-by: Casey Connolly <kcxt@postmarketos.org>