treewide: split chroots from workdir
Introduce a new "cache" subdirectory in the pmbootstrap workdir, all the
cache and config bits go in here, anything that needs to be accessible
from inside a chroot. The whole dir is then bind-mounted into the chroot
as /cache with appropriate symlinks.

This dir is in the config as config.cache.

The various cache_* dirs are renamed, and a new migration is added to
move the existing dirs.

Signed-off-by: Casey Connolly <kcxt@postmarketos.org>