chroot: formalise chroot API
We had a whole bunch of functions for doing things with chroots
which could leave them in weird states, and in general the API was
confusing. This refactor simplifies things down and stops exposing
the chroot.mount submodule externally. Now chroots can be set up with
pmb.chroot.init(Chroot) and torn down with pmb.chroot.shutdown(Chroot).
The state of the mountpoints, the "in-pmbootstrap" tag and the other
symlinks that must be created is handled entirely internally, other
modules should not try to use these APIs.

Signed-off-by: Casey Connolly <kcxt@postmarketos.org>