core: config: assume cache dir should be relative to work dir unless set explicitly
This is not the most elagent solution, but basically implement the logic
that if you specify a custom work dir but not a custom cache dir, it
should be a subdir named "cache".

Setting a custom cache dir may be useful for working with multiple
contexts (assuming a custom aports per context).

Signed-off-by: Casey Connolly <kcxt@postmarketos.org>