Load ./pmbootstrap.conf instead of global config if found
Since the work and cache dirs are now separated, it is easy for one to
have a custom pmbootstrap config which points the work dir to ./.pmb,
and a local aports checkout ./pmaports. Let's finish things off by
teaching pmbootstrap to look for a local config file to declare these
things.

We check the working directory and all parent directories, adjusting
os.cwd() when found.

This way, one can have a "project" directory like so:

$ ls -A | cat
total 16K
4.0K .pmb
4.0K firmware
4.0K pmaports
4.0K pmbootstrap.conf

With the pmbootstrap.conf containing

$ cat pmbootstrap.conf
[pmbootstrap]
aports = ./pmaports
work = ./.pmb
...

And now when pmbootstrap is invoked in this directory, it will use this
local config and workdir. However the cache remains global, so all the
cached apk packages, ccache, the local repo and signing keys are all
still in ~/.local/var/pmbootstrap/cache. This drastically reduces the
data duplication!

The local package repo is now also kept in the work dir, so it is
specific to each project directory.

Signed-off-by: Casey Connolly <kcxt@postmarketos.org>