CI: Skip removing files that start with test in docs check
Tests were moved out of the pmb directory into their own test directory,
so no need to try to exclude them when they shouldn't be there anyway.
If test "documentation" appears in the docs with this, there's a test
that is where it shouldn't be and a separate bug.

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2619