sideload: skip unnecessary apkindex updates
When sideloading a package apk on the device may attempt to update it's
APKINDEX cache, this can be annoying when e.g. sideloading a kernel on
first boot (or any other time the APKINDEX cache needs refreshing).

Let's avoid this case by running apk with --no-network if none of
the packages being sideloaded have runtime dependencies, since we can
be 100% confident in this case that apk doesn't need to update its
indexes.

Signed-off-by: Casey Connolly <kcxt@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2760>