pmbootstrap: fix running ssh in a user namespace
Since we are root in the namespace, ssh will attempt to use the root
users config, additionally the ssh_config.d directory will have invalid
permissions (since it's owned by root on the host which is not mapped
into the userns).

Hack around all of these with some fancy mounts in the sandbox.

The "proper" way to do this would be to stop using the host rootfs
alltogether and unshare ourselves into an Alpine chroot, with the
users .ssh directory mounted in. But that will require refactoring
pmbootstrap's init code to be able to parse args and do actions before
entering the namespace.

Signed-off-by: Casey Connolly <kcxt@postmarketos.org>