Fix crash in "zap -a" and "zap -o" with apk3
With apkv3, APKINDEX files get deleted when running "apk cache clean".
This is unexpected and needs to be investigated. Meanwhile run
cache_clean() on the native arch last, as it needs the APKINDEX for the
Alpine main repo of the native arch (even if cleaning other arches) in
order to set up the chroot where we run "apk cache clean". This prevents
crashing in "pmbootstrap zap -a" and "pmbootstrap zap -o".

Related: BPO issue 160
Related: pmb issue 2627
Part-of: https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2636