pmb: Add --shallow-initial-clone argument to init
In particular, this is useful for the postmarketOS os-installer "UI"
where we only care about the latest pmaports state and not the entire
history. Not having to download the entire history both saves time in
downloading the repository but also IO, especially since you typically
flash installers onto USB flash drives which can be terribly slow.

However, any other sort of installer script could probably find this
useful as well.