Revert "pmb.qemu: set default image size of 4G (!1886)"
The original commit that does this provides no explanation as to why
this was done, however digging through the comments the following
rationale can be found:

> @drebrez: @Minecrell What you think about adding a default value for the --image-size?

> @Minecrell: That's probably a good idea, but it might be a bit hard to choose a default image size.
>
> An alternative would be to use a larger image together with qcow (so it allocates space as needed), but setting that up would be require some more effort.

> @drebrez: A default image size that can contains at least the biggest UI package we have and some additional space, probably the plasma-mobile-extras.
>
> IIRC we already specify somewhere that pmbootstrap requires various GB of free space, so I don't think it will cause problems if we set a default of 2 or 3 or 5 GB.

> @drebrez: The plasma-mobile-extras UI with some other packages (chromium and some other utility) already uses more than 2G.
>
> My suggestion is to set a default of at least 3GB

> @Minecrell: I prefer powers of 2. I.e. 4 GiB :D

> @drebrez: I've set a default of 4 GB

So it seems the reasoning was to ensure that pmbootstrap's image size
always is big enough for large UIs. However in practice it seems to do
the opposite as I ran into problems with this default when trying to run
a RetroArch image in QEMU due to it being larger than the default value
for --image-size (4 GB).

Based on the qouted conversation, one would expect that just removing
the default would not help matters. Yet it does, and indeed fixes my
issue. So it seems something has changed in pmbootstrap between now and
then, or the change was misinformed from the beginning.

Either way, this is caused by that run() in pmb/qemu/run.py checks if
args.image_size is defined, and uses that to determine whether it should
resize the image or not. But since this change gives is a default value,
that always evaluates to true and so the code always ends up calling
resize_image(). That function, in turn, checks if the provided image
size is smaller than the image already is, and if so errors out with an
error message.

All good, except that nowadays images can easily be bigger than this
default and pmbootstrap seems to handle that just fine without an
explicit resize here (I ran pmbootstrap -y zap -a to verify this).
Furthermore, the error message seems to assume --image-size explicitly
was defined and isn't all that helpful if it was "implicitly defined"
through the default value.

So, with all that in mind, let's just revert this change.

This reverts commit a76897b4a19b50940d185d1a0ec72de2f7c76c47.