pmb.ci: Add new "with-dot-git" option
This can be used by CI scripts that need to access the git history.
Previously, this could be achieved by using the "native" option, but
it's not an ideal solution as it means the script has to rely on tooling
already available on the host, which is limiting and can be brittle. As
such, introduce this option that solves the aforementioned use-case
without the limitations of "native".