pmb.helpers.frontend: Deprecate 'pmbootstrap lint'
The issue below contains the full rationale. But in short, the 'lint'
subcommand was introduced before the 'ci' subcommand was created and
effectively implements the same concept of being able to run the CI
pipelines locally but hard-codes the apkbuild-lint task. Instead of
keeping this legacy code around, let's deprecate it and reimplement the
lint task in a way that's compatible with 'pmbootstrap ci'.

See https://gitlab.postmarketos.org/postmarketOS/pmaports/-/issues/3665