Fix /usr merge breaking stable branches
The /usr merge should only be done if it is supported by the branch, as
indicated in supported_usr_merge in pmaports.cfg. Currently it works in
edge, but not in v25.06 or older.

Without this patch, v25.06 installations currently fail when building
the initramfs, e.g. with qemu-amd64 and the console UI:

12:04:42.337639 hookfiles: unable to process hook file "/usr/share/mkinitfs/files-extra/00-initramfs-extra-base.files": unable to add "/usr/sbin/udevd": getFile: failed to stat file "/usr/sbin/udevd": stat /usr/sbin/udevd: no such file or directory (also tried "/usr/sbin/udevd.zst": stat /usr/sbin/udevd.zst: no such file or directory)

Fixes: b4338fd9 ("pmb.chroot.init: bootstrap the /usr-merge by creating symlinks")

Part-of: https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2687