pmb, test: Use git switch instead of git checkout
In these cases, we specifically want to switch branches, so git switch
does exactly what we want. While git checkout also works, it can do
other things too which potentially could lead to it doing the wrong
thing if an incorrect value is passed. It also makes it more clear what
the code is trying to do.