pmb: checksum: fail if verification fails
The abuild overrides had a verify() function that caused abuild to
only warn (and not fail) if running `pmbootstrap checksum --verify`.
If there's a mismatch, pmb should fail and not warn. The main reason
for this is the pmaports CI checksum test, if pmb doesn't fail when
checksums mismatch, then the test always passes even if checksums are
wrong.