kconfig: strip quotes when checking string values
This was incorrectly including explicit quotes when matching options
in the config, because previously this function only had to deal with
string values in kconfig  (e.g. `CONFIG_FOO="bar bazz"`). In order
to include `y` and `m`, which are unquoted strings, this removes the
explicit quotes from the regex (they should still match on `.*`), and
strips any quotes from the match group when compating to the expected
string.

fixes #2665