kconfig: refactor tristate value checking
This correctly handles all conditions now, whereas before it could print
"should be m when is y" if an option was =y but failed to parse for some
reason, or if the option was set to some invalid tristate value. I think
this is also a bit more cleaner to read, and avoids 1 extra call to
is_set_str. A future improvement might be to stop abusing is_set_str to
tease out the value of a given option in the config 😅

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2698>