pmb.config.init: Gracefully handle there not being any providrs for _pmb_select
As reported in the merge request removing rproc selection for
soc-qcom-msm8916[1], in the case of `_pmb_select` being set to a name
without providers, `pmb init` will get stuck on the provider selection
step. This is obviously not good as it means the device cannot be
selected at all for an unintuitive reason.

So, instead just skip the provider selection and print a warning if
there are no providers for a given name. While bugs like this of course
should be fixed on a packaging level, let's handle it gracefully as to
not completely block people's workflows.

 [1]: https://gitlab.postmarketos.org/postmarketOS/pmaports/-/merge_requests/7075#note_505426