docs: remove sphinxext.opengraph
This extension generates open graph preview images for sharing on social
media. While this is nice in theory, this has some drawbacks:

* Currently the preview doesn't work for pmbootstrap ("<meta
  property="og:image…" from the source code points to a non-existing
  URL, not sure why...).

* Time for building the site and resulting file size get bigger when the
  extension is enabled and configured to make social graph images.
  Currently, when building docs.postmarketos.org with index,
  policies-and-processes and dpo:
    without opengraph:
      2.6s, 2.3M
    with opengraph:
      4.6s, 3.9M
  This is already an obvious difference, and it will diverge much
  further as we add more documentation (which we plan to do).

* The preview images can't display SVG logos, so we currently have no
  logo there. It also may not be able to display some characters we use
  like emojis and may print additional warnings during the build.

I had only added this because it was also in the GNOME handbook, which
I've used as starting point for docs.postmarketos.org. IMHO it is not
worth it, so let's remove it.

Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2712>