pmb, test: Format docstrings consistently with D2 Ruff rules
As part of fixing the related issue, I want to start with making sure
our docstrings are formatted consistently before adding more. The D2
rules in Ruff concern stylistic formatting and as usual with stylistic
things are to some degree subjective. In particular, some of them are
mutually exclusive. This commit chooses D211 over D203 and D213 over
D212 due to personal preference, but there's even more ones that could
be disabled if we want to. Either way, something has to be choosen so I
went with enabling everything except for where a choice was necessary.

Related: https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/issues/2351
Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>