pmb, test: Format docstrings consistently with D2 Ruff rules
The D2 rules in Ruff concern stylistic formatting and as usual with
stylistic things are to some degree subjective. In particular, some of
them are mutually exclusive. This commit chooses the "pep257" style as
this was preferred in a brief discussion[1].

 [1]: https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2741#note_514833

Related: https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/issues/2351
Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2741>