pmb.build: Fix bootstrap of systemd repo
systemd-stage0 depends on dbus-dev<9990 and pmbootstrap would correctly
figure out that this package should use the dbus package from Alpine for
building, but since dbus-dev is also part of pmaports, it was also in
all_pkgnames. This would then result in pmbootstrap thinking it would
need to build the dbus from Alpine and incorrectly assuming there was a
cyclic dependency. We can fix this by only applying this logic for
dependencies that are being taken from pmaports, not those from aports.

I can successfully select qemu-loongarch64 in pmb init, enable systemd
and then build an image for it with this, prior it would fail.

Signed-off-by: Jens Reidel <adrian@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2747>