pmb.chroot.apk: Reduce repetition in apk package check
This is more Pythonic by virtue of being consistent with the
Python principle of that it's easier to ask for forgiveness than
permission. But, it also reduces the number of times we have to
repeat the package name of apk by one.

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>
Part-of: <https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/merge_requests/2763>