pmb: Check for expected page size
There is a bug somewhere that causes issues when the page size is
not 4096. Likely in qemu-user. Print a warning for now, might help
someone. Only reason I'm choosing to just warn instead of erroring
out is because not all commands rely on QEMU so pmbootstrap might
still be useful in some way even with these problems.

See https://gitlab.postmarketos.org/postmarketOS/pmbootstrap/-/issues/2690

Signed-off-by: Stefan Hansson <newbyte@postmarketos.org>